+++ src/main/java/com/org/bgv/company/dto/EmployerPackageConfigPreviewResponseDTO.java
@@ -1,4 +1,5 @@
 package com.org.bgv.company.dto;
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -18,9 +19,9 @@ public class EmployerPackageConfigPreviewResponseDTO {
     private Long employerPackageId;
     private String packageName;
 
-    private Double basePrice;
-    private Double addonPrice;
-    private Double estimatedTotalPrice;
+    private BigDecimal basePrice;
+    private BigDecimal addonPrice;
+    private BigDecimal estimatedTotalPrice;
     private Boolean finalPriceDynamic;
 
     private List<CategoryPreviewDTO> categories;
